Added an exception to allow less and more to work
The UNIX tools 'less' and 'more' require the ability to write to
a file named /usr/bin/.sysless. In order to prohibit these tools
from printing an error message, we add a temporary exception in
the validity checking code to allow the operation to proceed even
though it is in a protected folder.